Updated io-factory calls through blkdev for 6.5 kernel
Added support to pass a value for the new blk_holder_ops argument that
is used by blkdev_get_by_dev() and blkdev_get_by_path(). This was
originally introduced in the linux kernel via commit
0718afd47f70cf46877c39c25d06b786e1a3f36c.

Also adapted the argument preferneces for blkdev_put() that uses the
holder as an indication for exclusive opens.  This was originally
introduced in the linux kernel via commit
2736e8eeb0ccdc71d1f4256c9c9a28f58cc43307.

Signed-off-by: Andrew Walsh <awalsh@redhat.com>